receipt-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
Pillow>=10.0.0
pydantic==2.9.0
requests==2.32.0